Reworked uncoated areas generation setup.
It turned out that PhysX doesn't support multiple materials applied to a single collider with UsdGeomSubsets. Initally I thought to use this setup, but I had to separate the inner panel faces and remain parts of the cap model. I also can't turn this panel into a collider now because Raycast used in ProductSensor will detect it as a separate collider. By now it is acceptable though, and I can do UV-unwrapping much easier.