Add a cap spawner with ovality and inclusion defects generation
When I decided to implement the cap spawner as an OmniGraph node, I though it will be much easier: nodes already have UI for inputs and you haven't to subscribe on some callbacks to execute their code. Things went well until I tried to use Replicator's functions inside the node. It just doesn't work properly, and nobody has told about it in the Internet. I wasted an entire day to unsuccessfully mess with it just to rewrite it in the main extension file in the end.